Added New String Function, Merged Into Larger File
Added atlasui_string_check, which is similar to strpos / strstr in that
it checks for a needle in a haystack. Its an alternative to strpos as it
accepts arrays, however only returns true or false (true if the needle
is in the haystack). Merged atlasui_string_clean and
atlasui_string_check into a strings file.